+++ opal/caregivers/api/views.py
@@ -1,18 +1,35 @@
 """This module is an API view that returns the encryption value required to handle listener's registration requests."""
+from typing import Any
 
+from django.conf import settings
+from django.core.mail import send_mail
 from django.db.models.functions import SHA512
+from django.db.models.query import QuerySet
+from django.template.loader import render_to_string
+from django.utils import timezone
+from django.utils.translation import gettext_lazy as _
 
+from rest_framework import serializers as drf_serializers
 from rest_framework import status
-from rest_framework.generics import RetrieveAPIView
+from rest_framework.generics import RetrieveAPIView, get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from opal.caregivers.api.serializers import RegistrationEncryptionInfoSerializer
-from opal.caregivers.models import RegistrationCode, RegistrationCodeStatus
+from opal.caregivers.api.mixins.put_as_create import AllowPUTAsCreateMixin
+from opal.caregivers.api.serializers import (
+    EmailVerificationSerializer,
+    RegistrationEncryptionInfoSerializer,
+    UpdateDeviceSerializer,
+)
+from opal.caregivers.models import Device, EmailVerification, RegistrationCode, RegistrationCodeStatus
+from opal.core.utils import generate_random_number
 from opal.patients.api.serializers import CaregiverPatientSerializer
 from opal.patients.models import Relationship
+from opal.users.models import User
+
+from .. import constants
 
 
 class GetRegistrationEncryptionInfoView(RetrieveAPIView):
@@ -29,6 +46,53 @@ class GetRegistrationEncryptionInfoView(RetrieveAPIView):
     serializer_class = RegistrationEncryptionInfoSerializer
     lookup_url_kwarg = 'hash'
     lookup_field = 'code_sha512'
+
+
+class UpdateDeviceView(AllowPUTAsCreateMixin):
+    """Class handling requests for updates or creations of device ids."""
+
+    permission_classes = [IsAuthenticated]
+    serializer_class = UpdateDeviceSerializer
+    lookup_url_kwarg = 'device_id'
+    lookup_field = 'device_id'
+
+    def put(self, request: Request, *args: Any, **kwargs: Any) -> Any:
+        """Handle incoming put request and redirect to update method.
+
+        Args:
+            request (Request): request object with parameters to update or create
+            args (Any): varied amount of non-keyworded arguments
+            kwargs (Any): varied amount of keyworded arguments
+
+        Returns:
+            HTTP `Response` success or failure
+        """
+        if self.request.method == 'PUT':
+            return self.update(request, *args, **kwargs)
+        return Response({'status': status.HTTP_404_NOT_FOUND})
+
+    def patch(self, request: Request, *args: Any, **kwargs: Any) -> Any:
+        """Handle incoming path request and redirect to partial update method.
+
+        Args:
+            request (Request): request object with parameters to update or create
+            args (Any): varied amount of non-keyworded arguments
+            kwargs (Any): varied amount of keyworded arguments
+
+        Returns:
+            HTTP `Response` success or failure
+        """
+        if self.request.method == 'PATCH':
+            return self.partial_update(request, *args, **kwargs)
+        return Response({'status': status.HTTP_404_NOT_FOUND})
+
+    def get_queryset(self) -> QuerySet[Device]:
+        """Provide the desired object or fails with 404 error.
+
+        Returns:
+            Device object or 404.
+        """
+        return Device.objects.filter(device_id=self.kwargs['device_id'])
 
 
 class GetCaregiverPatientsList(APIView):
@@ -56,3 +120,167 @@ class GetCaregiverPatientsList(APIView):
             response = Response([], status=status.HTTP_400_BAD_REQUEST)
 
         return response
+
+
+class RetrieveRegistrationCodeMixin(APIView):
+    """Mixin class that provides `get_queryset()` to lookup a `RegistrationCode` based on a given `code`."""
+
+    def get_queryset(self) -> QuerySet[RegistrationCode]:
+        """
+        Override get_queryset to filter RegistrationCode by registration code.
+
+        Returns:
+            The queryset of RegistrationCode
+        """
+        code = self.kwargs.get('code') if hasattr(self, 'kwargs') else None
+        return RegistrationCode.objects.select_related(
+            'relationship__caregiver',
+            'relationship__caregiver__user',
+        ).prefetch_related(
+            'relationship__caregiver__email_verifications',
+        ).filter(code=code, status=RegistrationCodeStatus.NEW)
+
+
+class VerifyEmailView(RetrieveRegistrationCodeMixin, APIView):
+    """View that initiates email verification for a given email address.
+
+    And send email to the user with the verification code.
+    """
+
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request: Request, code: str) -> Response:  # noqa: WPS210
+        """
+        Generate a random verification code and set up the EmailVerification instance.
+
+        If the user requested to re-send the code too soon, it fails.
+
+        Args:
+            request: the HTTP request.
+            code: registration code.
+
+        Raises:
+            ValidationError: if re-sending the verification code was requested too soon.
+
+        Returns:
+            Http response with empty message.
+        """
+        registration_code = get_object_or_404(self.get_queryset())
+
+        input_serializer = EmailVerificationSerializer(data=request.data, fields=('email',))
+        input_serializer.is_valid(raise_exception=True)
+
+        email = input_serializer.validated_data['email']
+        verification_code = generate_random_number(constants.VERIFICATION_CODE_LENGTH)
+        caregiver = registration_code.relationship.caregiver
+        try:
+            email_verification = registration_code.relationship.caregiver.email_verifications.get(
+                email=email,
+            )
+        except EmailVerification.DoesNotExist:
+            email_verification = EmailVerification.objects.create(
+                caregiver=caregiver,
+                code=verification_code,
+                email=email,
+                sent_at=timezone.now(),
+            )
+            self._send_email(email_verification, caregiver.user, request.LANGUAGE_CODE)
+        else:
+            # in case there is an error sent_at is None, but wont happen in fact
+            time_delta = timezone.now() - timezone.localtime(email_verification.sent_at)
+            if time_delta.total_seconds() >= constants.TIME_DELAY:
+                input_serializer.update(
+                    email_verification,
+                    {
+                        'code': verification_code,
+                        'is_verified': False,
+                        'sent_at': timezone.now(),
+                    },
+                )
+                self._send_email(email_verification, caregiver.user, request.LANGUAGE_CODE)
+            else:
+                raise drf_serializers.ValidationError(
+                    _('Please wait 10 seconds before requesting a new verification code.'),
+                )
+
+        return Response()
+
+    def _send_email(  # noqa: WPS210
+        self,
+        email_verification: EmailVerification,
+        user: User,
+        language: str,
+    ) -> None:
+        """
+        Send verification email to the user with an template according to the user language.
+
+        Args:
+            email_verification: object EmailVerification.
+            user: object User.
+            language: language code from the request data.
+        """
+        email_subject = _('Opal Verification Code')
+
+        template_plain = 'email/verification_code.txt'
+        parameters = {
+            'code': email_verification.code,
+            'first_name': user.first_name,
+            'last_name': user.last_name,
+        }
+        msg_plain = render_to_string(
+            template_plain,
+            parameters,
+        )
+        template_html = 'email/verification_code.html'
+        msg_html = render_to_string(
+            template_html,
+            parameters,
+        )
+        send_mail(
+            email_subject,
+            msg_plain,
+            settings.EMAIL_HOST_USER,
+            [email_verification.email],
+            html_message=msg_html,
+        )
+
+
+class VerifyEmailCodeView(RetrieveRegistrationCodeMixin, APIView):
+    """View that verifies the user-provided verification code with the actual one."""
+
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request: Request, code: str) -> Response:  # noqa: WPS210
+        """
+        Verify that the provided code matches the expected one.
+
+        And if so, it deletes the EmailVerification,
+        and updates the email on the user instance with the verified one.
+
+        Args:
+            request: the HTTP request.
+            code: registration code.
+
+        Returns:
+            Http response with empty message.
+        """
+        registration_code = get_object_or_404(self.get_queryset())
+
+        input_serializer = EmailVerificationSerializer(data=request.data, fields=('code', 'email'))
+        input_serializer.is_valid(raise_exception=True)
+
+        verification_code = input_serializer.validated_data['code']
+        email = input_serializer.validated_data['email']
+
+        email_verification = get_object_or_404(
+            registration_code.relationship.caregiver.email_verifications,
+            code=verification_code,
+            email=email,
+        )
+
+        email_verification.delete()
+        user = registration_code.relationship.caregiver.user
+        user.email = email_verification.email
+        user.save()
+
+        return Response()