+++ opal/caregivers/models.py
@@ -27,6 +27,7 @@ class CaregiverProfile(models.Model):
     legacy_id = models.PositiveIntegerField(
         verbose_name=_('Legacy ID'),
         validators=[MinValueValidator(1)],
+        unique=True,
         null=True,
         blank=True,
     )
@@ -136,6 +137,17 @@ class Device(models.Model):
     is_trusted = models.BooleanField(
         verbose_name=_('Trusted Device'),
         default=False,
+    )
+
+    push_token = models.CharField(
+        verbose_name=_('Push Token'),
+        max_length=256,
+        blank=True,
+    )
+
+    modified = models.DateTimeField(
+        verbose_name=_('Last Modified'),
+        auto_now=True,
     )
 
     class Meta: